Cleaning up organization and namespacing
