--- Nav2SLAMExampleProject/Assets/Scripts/FreeCam.cs
@@ -1,124 +1,0 @@
-using System;
-using UnityEngine;
-
-/// <summary>
-///     A simple free camera to be added to a Unity game object.
-///     Keys:
-///     wasd / arrows	- movement
-///     q/e 			- down/up (local space)
-///     r/f 			- up/down (world space)
-///     pageup/pagedown	- up/down (world space)
-///     hold shift		- enable fast movement mode
-///     right mouse  	- enable free look
-///     mouse			- free look / rotation
-/// </summary>
-public class FreeCam : MonoBehaviour
-{
-    /// <summary>
-    ///     Normal speed of camera movement.
-    /// </summary>
-    [SerializeField]
-    float m_MovementSpeed = 10f;
-
-    /// <summary>
-    ///     Speed of camera movement when shift is held down,
-    /// </summary>
-    [SerializeField]
-    float m_FastMovementSpeed = 100f;
-
-    /// <summary>
-    ///     Sensitivity for free look.
-    /// </summary>
-    [SerializeField]
-    float m_FreeLookSensitivity = 3f;
-
-    /// <summary>
-    ///     Amount to zoom the camera when using the mouse wheel.
-    /// </summary>
-    [SerializeField]
-    float m_ZoomSensitivity = 10f;
-
-    /// <summary>
-    ///     Amount to zoom the camera when using the mouse wheel (fast mode).
-    /// </summary>
-    [SerializeField]
-    float m_FastZoomSensitivity = 50f;
-
-    /// <summary>
-    ///     Set to true when free looking (on right mouse button).
-    /// </summary>
-    bool m_Looking;
-
-    void Update()
-    {
-        var fastMode = Input.GetKey(KeyCode.LeftShift) || Input.GetKey(KeyCode.RightShift);
-        var movementSpeed = fastMode ? m_FastMovementSpeed : m_MovementSpeed;
-
-        if (Input.GetKey(KeyCode.A) || Input.GetKey(KeyCode.LeftArrow))
-            transform.position = transform.position + -transform.right * movementSpeed * Time.deltaTime;
-
-        if (Input.GetKey(KeyCode.D) || Input.GetKey(KeyCode.RightArrow))
-            transform.position = transform.position + transform.right * movementSpeed * Time.deltaTime;
-
-        if (Input.GetKey(KeyCode.W) || Input.GetKey(KeyCode.UpArrow))
-            transform.position = transform.position + transform.forward * movementSpeed * Time.deltaTime;
-
-        if (Input.GetKey(KeyCode.S) || Input.GetKey(KeyCode.DownArrow))
-            transform.position = transform.position + -transform.forward * movementSpeed * Time.deltaTime;
-
-        if (Input.GetKey(KeyCode.Q))
-            transform.position = transform.position + -transform.up * movementSpeed * Time.deltaTime;
-
-        if (Input.GetKey(KeyCode.E))
-            transform.position = transform.position + transform.up * movementSpeed * Time.deltaTime;
-
-        if (Input.GetKey(KeyCode.R) || Input.GetKey(KeyCode.PageUp))
-            transform.position = transform.position + Vector3.up * movementSpeed * Time.deltaTime;
-
-        if (Input.GetKey(KeyCode.F) || Input.GetKey(KeyCode.PageDown))
-            transform.position = transform.position + -Vector3.up * movementSpeed * Time.deltaTime;
-
-        if (m_Looking)
-        {
-            var newRotationX = transform.localEulerAngles.y + Input.GetAxis("Mouse X") * m_FreeLookSensitivity;
-            var newRotationY = transform.localEulerAngles.x - Input.GetAxis("Mouse Y") * m_FreeLookSensitivity;
-            transform.localEulerAngles = new Vector3(newRotationY, newRotationX, 0f);
-        }
-
-        var axis = Input.GetAxis("Mouse ScrollWheel");
-        if (axis != 0)
-        {
-            var zoomSensitivity = fastMode ? m_FastZoomSensitivity : m_ZoomSensitivity;
-            transform.position = transform.position + transform.forward * axis * zoomSensitivity;
-        }
-
-        if (Input.GetKeyDown(KeyCode.Mouse1))
-            StartLooking();
-        else if (Input.GetKeyUp(KeyCode.Mouse1)) StopLooking();
-    }
-
-    void OnDisable()
-    {
-        StopLooking();
-    }
-
-    /// <summary>
-    ///     Enable free looking.
-    /// </summary>
-    void StartLooking()
-    {
-        m_Looking = true;
-        Cursor.visible = false;
-        Cursor.lockState = CursorLockMode.Locked;
-    }
-
-    /// <summary>
-    ///     Disable free looking.
-    /// </summary>
-    void StopLooking()
-    {
-        m_Looking = false;
-        Cursor.visible = true;
-        Cursor.lockState = CursorLockMode.None;
-    }
-}
