+++ Nav2SLAMExampleProject/Assets/Scripts/RosInterfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7cdf87e755340954b8e7ce5added780e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
