+++ Nav2SLAMExampleProject/Assets/Scripts/Utility/FreeCam.cs.meta
