+++ Nav2SLAMExampleProject/Packages/manifest.json
@@ -10,6 +10,7 @@
     "com.unity.robotics.visualizations": "https://github.com/Unity-Technologies/ROS-TCP-Connector.git?path=/com.unity.robotics.visualizations#v0.6.0",
     "com.unity.robotics.warehouse": "https://github.com/Unity-Technologies/Robotics-Warehouse.git?path=com.unity.robotics.warehouse#nav2-example",
     "com.unity.test-framework": "1.1.24",
+    "com.unity.testtools.codecoverage": "1.1.0",
     "com.unity.textmeshpro": "3.0.6",
     "com.unity.timeline": "1.4.8",
     "com.unity.ugui": "1.0.0",
